+++ redus_dart/lib/src/reactivity/advanced/custom_ref.dart
@@ -78,5 +78,8 @@ class _CustomRef<T> implements Ref<T> {
   }
 
   @override
+  T call() => value;
+
+  @override
   String toString() => 'CustomRef($value)';
 }